% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Covmeta.R
\name{cov_meta}
\alias{cov_meta}
\title{overall covariance of random effect meta-analysis in presence of interaction}
\usage{
cov_meta(
  b1 = NULL,
  v1 = NULL,
  b2 = NULL,
  v2 = NULL,
  cov_b1b2 = NULL,
  sample_size = NULL,
  effect = "random",
  ...
)
}
\arguments{
\item{b1}{the coefficient of main effect from the regression}

\item{v1}{the variance of main effect from the regression}

\item{b2}{the coefficient of the interaction effect from the regression}

\item{v2}{the variance of interaction effect from the regression}

\item{cov_b1b2}{the covariance of main effect and interaction effect from the regression}

\item{sample_size}{the sample size of the study}

\item{effect}{fixed or random}
}
\value{
overall main effect, overall interaction effect and overall covariance
}
\description{
overall covariance of random effect meta-analysis in presence of interaction
}
\examples{
#read in datatable
studies<-read.csv('https://raw.githubusercontent.com/enwuliu/meta-analysis/main/random_effect_meta_sim.csv',header=T)
#function arguments
b1<-studies$b1; #beta1 of the 10 cohorts
v1<-studies$var_b1; #variances of the coefficients
b2<-studies$b2; #beta1 of the 10 cohorts
v2<-studies$var_b2; #variances of the coefficients
cov_b1b2<-studies$cov_b1b2; #covariance of b1 and b2
sample_size<-studies$sample_size; #sample size
cov_meta(b1,v1,b2,v2, cov_b1b2, sample_size,'fixed')
cov_meta(b1,v1,b2,v2, cov_b1b2, sample_size,'random')
}
